Move PackageManager where they can be reused
JIRA-Ref: CMK-15344
